removed view builder and removed specific if statements for views
