+++ RedSolApp/RedSolApp/TabPages/MyResourcesView/OrganizationCardView.swift
